Initial attempt at swagger-ui-express
This adds swagger-ui-express and jsdoc annotations.
Not having much luck with defining the application/json
body for /user/login, which seems to be additionally
complicated in openapi 3.0 (previously would have been
just formData parameter)

As swagger-ui-express can simply take an openapi definition
json as input, and Maria is already working on swagger definition
for the Python backend, we could reuse that, rather than doing
extra work to re-generate it here.